requirements.txt
arrow==1.2.2
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
boto3==1.20.20
botocore==1.23.20
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
cssselect2==0.5.0
daphne==3.0.2
Django==3.2.9
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-filter==21.1
django-phonenumber-field==6.0.0
django-renderpdf==3.0.1
django-ses==2.3.1
django-storages==1.12.3
django-test-plus==2.2.0
django-timezone-field==4.2.1
django-uwsgi==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==13.3.4
fonttools==4.31.2
future==0.18.2
gobject==0.1.0
gunicorn==20.1.0
hashids==1.3.1
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
iniconfig==1.1.1
jmespath==0.10.0
Markdown==3.3.6
# mysqlclient==2.1.0
packaging==21.3
phonenumbers==8.12.37
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyphen==0.12.0
PySocks==1.7.1
pytest==7.1.1
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.20.0
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
termcolor==1.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
twilio==6.0.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.7
weasyprint==54.2
webencodings==0.5.1
zipp==3.6.0
zoomus==1.1.6
zope.interface==5.4.0
zopfli==0.2.1
stripe==3.3.0
